refactor: getDirectories function into one file
