+++ scripts/utils/fontlist-generator.js
@@ -2,14 +2,7 @@ const _ = require("lodash")
 const fs = require("fs-extra")
 const jsonfile = require("jsonfile")
 
-// Find names of all packages.
-const getDirectories = () =>
-  fs
-    .readdirSync("./packages", { withFileTypes: true })
-    .filter(dirent => dirent.isDirectory())
-    .map(dirent => dirent.name)
-
-const directories = getDirectories()
+const { directories } = require("./utils")
 
 const fontlist = []
 const league = []
